+++ src/Polly.Net40Async/Properties/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 ﻿using System;
 using System.Reflection;
 [assembly: AssemblyTitle("Polly")]
-[assembly: CLSCompliant(true)]
+[assembly: CLSCompliant(false)] // Because Nito.AsycEx, on which Polly.Net40Async depends, is not CLSCompliant.